Hook up Open and Create buttons in the PR list
