+++ src/GitHub.App/SampleData/PullRequestCreationViewModelDesigner.cs
@@ -1,0 +1,14 @@
+using System.Diagnostics.CodeAnalysis;
+using GitHub.ViewModels;
+
+namespace GitHub.SampleData
+{
+    [ExcludeFromCodeCoverage]
+    public class PullRequestCreationViewModelDesigner : BaseViewModelDesigner, IPullRequestCreationViewModel
+    {
+        public PullRequestCreationViewModelDesigner()
+        {
+
+        }
+    }
+}
