+++ src/GitHub.App/SampleData/PullRequestDetailViewModelDesigner.cs
@@ -1,0 +1,14 @@
+using System.Diagnostics.CodeAnalysis;
+using GitHub.ViewModels;
+
+namespace GitHub.SampleData
+{
+    [ExcludeFromCodeCoverage]
+    public class PullRequestDetailViewModelDesigner : BaseViewModelDesigner, IPullRequestDetailViewModel
+    {
+        public PullRequestDetailViewModelDesigner()
+        {
+
+        }
+    }
+}
