+++ src/GitHub.App/SampleData/SampleViewModels.cs
@@ -17,8 +17,6 @@ using GitHub.VisualStudio.TeamExplorer.Home;
 using ReactiveUI;
 using GitHub.VisualStudio.TeamExplorer.Connect;
 using System.Collections.ObjectModel;
-using GitHub.Collections;
-using System.Reactive.Linq;
 using System.Linq;
 
 namespace GitHub.SampleData
@@ -29,22 +27,9 @@ namespace GitHub.SampleData
         public ICommand Cancel { get; set; }
         public bool IsShowing { get; set; }
         public string Title { get; set; }
-    }
 
-    [ExcludeFromCodeCoverage]
-    public class PullRequestListViewModelDesigner : BaseViewModelDesigner, IPullRequestListViewModel
-    {
-        public PullRequestListViewModelDesigner()
-        {
-            var prs = new TrackingCollection<IPullRequestModel>(Observable.Empty<IPullRequestModel>());
-            prs.Add(new PullRequestModel(399, "Let's try doing this differently", new AccountDesigner { Login = "shana", IsUser = true }, DateTimeOffset.Now - TimeSpan.FromDays(1)));
-            prs.Add(new PullRequestModel(389, "Build system upgrade", new AccountDesigner { Login = "haacked", IsUser = true }, DateTimeOffset.Now - TimeSpan.FromMinutes(2)) { CommentCount = 4, HasNewComments = false });
-            prs.Add(new PullRequestModel(409, "Fix publish button style", new AccountDesigner { Login = "shana", IsUser = false }, DateTimeOffset.Now - TimeSpan.FromHours(5)) { CommentCount = 27, HasNewComments = true });
-            PullRequests = prs;
-        }
-
-        public ITrackingCollection<IPullRequestModel> PullRequests { get; set; }
-        public IPullRequestModel SelectedPullRequest { get; set; }
+        public void Initialize(object data)
+        { }
     }
 
     [ExcludeFromCodeCoverage]
