+++ src/GitHub.VisualStudio/Styles/ActionLinkButton.xaml
@@ -11,7 +11,9 @@
       <Setter.Value>
         <ControlTemplate TargetType="Button">
           <TextBlock>
-            <Hyperlink Foreground="{TemplateBinding Foreground}">
+            <Hyperlink Command="{TemplateBinding Command}"
+                       CommandParameter="{TemplateBinding CommandParameter}"
+                       Foreground="{TemplateBinding Foreground}">
               <Hyperlink.Resources>
                 <Style TargetType="{x:Type Hyperlink}">
                   <Style.Triggers>
@@ -62,7 +64,9 @@
       <Setter.Value>
         <ControlTemplate TargetType="Button">
           <TextBlock>
-            <Hyperlink Foreground="{TemplateBinding Foreground}">
+            <Hyperlink Command="{TemplateBinding Command}"
+                       CommandParameter="{TemplateBinding CommandParameter}"
+                       Foreground="{TemplateBinding Foreground}">
               <Hyperlink.Resources>
                 <Style TargetType="{x:Type Hyperlink}">
                   <Style.Triggers>
