+++ src/GitHub.VisualStudio/UI/Views/Controls/PullRequestListItem.xaml
@@ -2,7 +2,9 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
                     xmlns:i18n="clr-namespace:GitHub.VisualStudio"
+                    xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views"
                     xmlns:models="clr-namespace:GitHub.Models;assembly=GitHub.Exports"
+                    xmlns:viewmodels="clr-namespace:GitHub.ViewModels;assembly=GitHub.Exports.Reactive"
                     xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI">
 
   <ResourceDictionary.MergedDictionaries>
@@ -105,9 +107,10 @@
           <ColumnDefinition Width="Auto" />
           <ColumnDefinition Width="*" />
         </Grid.ColumnDefinitions>
-        <Button x:Name="cloneLink"
+        <Button x:Name="prHashtagLink"
                 Grid.Column="0"
-                Command="{Binding OpenPR}"
+                Command="{Binding OpenPR, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:PullRequestListView}}}"
+                CommandParameter="{Binding Number}"
                 Content="{Binding Number}"
                 FontFamily="Segoe UI"
                 FontSize="11"
