+++ src/GitHub.VisualStudio/UI/Views/PullRequestCreationView.xaml
@@ -1,0 +1,32 @@
+﻿<local:GenericPullRequestCreationView x:Class="GitHub.VisualStudio.UI.Views.PullRequestCreationView"
+                                      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                                      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                                      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                                      xmlns:helpers="clr-namespace:GitHub.VisualStudio.Helpers"
+                                      xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views"
+                                      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                                      xmlns:sampleData="clr-namespace:GitHub.SampleData;assembly=GitHub.App"
+                                  Background="{DynamicResource GitHubVsToolWindowBackground}"
+                                  DataContext="{Binding ViewModel}"
+                                      d:DesignHeight="300"
+                                      d:DesignWidth="300"
+                                      mc:Ignorable="d">
+
+  <d:DesignProperties.DataContext>
+    <Binding>
+      <Binding.Source>
+        <sampleData:PullRequestCreationViewModelDesigner />
+      </Binding.Source>
+    </Binding>
+  </d:DesignProperties.DataContext>
+
+  <UserControl.Resources>
+    <ResourceDictionary>
+      <ResourceDictionary.MergedDictionaries>
+        <helpers:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/SharedDictionary.xaml" />
+        <helpers:SharedDictionaryManager Source="pack://application:,,,/GitHub.UI;component/SharedDictionary.xaml" />
+      </ResourceDictionary.MergedDictionaries>
+    </ResourceDictionary>
+  </UserControl.Resources>
+  <Grid />
+</local:GenericPullRequestCreationView>
