+++ src/GitHub.VisualStudio/UI/Views/PullRequestListView.xaml
@@ -58,6 +58,24 @@
           </Setter.Value>
         </Setter>
       </Style>
+      
+      <Style x:Key="MenuItemBaseStyle" TargetType="MenuItem">
+        <Style.Triggers>
+            <Trigger Property="IsMouseOver" Value="true">
+                <Setter Property="Background" Value="#0a99f3" />
+                <Setter Property="Foreground" Value="White"/>
+            </Trigger>
+        </Style.Triggers>
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type MenuItem}">
+                    <Grid Background="{TemplateBinding Background}">
+                        <MenuItem Header="{TemplateBinding Header}" />
+                    </Grid>
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
     </ResourceDictionary>
 
   </UserControl.Resources>
@@ -118,7 +136,7 @@
 
       </WrapPanel>
 
-      <ui:GitHubActionLink Grid.Column="1" HorizontalAlignment="Right" Content="Create New" />
+      <ui:GitHubActionLink Grid.Column="1" HorizontalAlignment="Right" Content="Create New" Command="{Binding CreatePR, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:PullRequestListView}}}"/>
 
     </Grid>
